Exploring windows support: get cake build working.
This gets the windows build of Docco to run over itself, and fixes `cake
build`

Now with 100% more just-hack-it.

The 'spawn' change in Cakefile is because of a windows specific nodejs
limitation: CreateProcessW only works on executable files, so any batch
files or other scripts cannot be launched using it.