Add common LaTeX file formats
Adds `dtx', 'cls', and `sty' files as LaTeX documents. DTX files are for documented sources and can be used to generate LaTeX class and package source files. `sty' and `cls' files are for packages and classes and can be generated by dtx files or written on their own. These files are the most likely use of docco in the TeX world.