add "variables" to defaults
Configuration is 'picky' so only parameters specified in defaults can be injected to configuration. There is no way to set custom variables to templates when you are running docco from grunt (like with grunt-docco). In this way you can pass arguments (for example the whole package.json) to the custom template, and you can access project specific attributes and embed them in the template.