Fix the install target
This commit fixes two problems related to cake install.

Firstly it was assumed that the prefix directory contains a
subdirectory named bin. When it didn't exist the script failed when
creating a symlink to docco. The problem is fixed by creating the bin
subdirectory.

Secondly an attempt to install (i.e. cp -rf) a directory named lib was
made. However the project doesn't contain such subdirectory. It was
removed from the list of directories to copy.