Fix the -m/--marked option
The argument to -m/--marked is a JSON file name containing additional
optins that should be passed to Marked. The problem was that the
default value of the command line option, as specified in the defaults
objects, was set to an empty object which definitely isn't a valid file
name. The result was an attempt to fs.readFileSync({}) which obviously
raised an exception.

This commit fixes the problem by setting the default value of
-m/--marked to null.