Add multi-line support.
This commit adds support for multi-line comments for languages that
use heredoc-style multi-line comments.  This change only enables
this feature for coffeescript and python, but others could be
added if they follow a similar syntax.  Notably, this change does
not add support for C-style comments which are more difficult to
support in the general case.

Derived from https://github.com/jashkenas/docco/pull/72