The rocco and shocco projects are broken for a number of years. Both seem abandoned.
Please remove the references to the rocco and shocco projects. Both are currently broken and their respective projects apparently abandoned.  At a minimum both seem to generate docs with hardcoded paths to CSS files which point to this repository.  Paths which have changed and broken over the years.  Both have primary homepages which are to act as demos, and both are unstyled for years as a result.

I have also filed an issue on the rocco project to ask that the project be marked as no longer supported.

https://github.com/rtomayko/rocco/issues/107

If it were not for the links on this docco main site I would not have gone down that path and wasted some time in the process of discovery as to why it was broken.

Thanks!