Added ability to ignore syntax highlight errors.
Highlight.js lacks support for languages like jade, so to avoid the parser throwing an error we just ignore the error and handle it gracefully. The rationale behind this is that the template can implement a client side library (prism.js or another) or a custom process for handling unsupported languages while still maintaining the docco file generation pipeline.